Phase 5.1: IMPORTANT Quality & Resilience Improvements
TIMEOUT WRAPAROUND SAFETY (Handles millis() rollover every ~49 days):
- timeout_manager.cpp: Fixed timeoutCheck(), timeoutElapsed(), timeoutRemaining()
- motion_control.cpp: Fixed all state machine timeout comparisons (6 instances)
- safety.cpp: Fixed stall check and alarm duration calculations
- config_unified.cpp: Fixed NVS save interval check
- Pattern: Changed `(now - last)` to `(uint32_t)(now - last)` for wraparound safety
- Ensures timeouts work correctly even across millis() overflow

FAULT LOG RING BUFFER FALLBACK:
- Added 8-entry circular buffer when queue is full (prevents silent data loss)
- fault_logging.cpp: Implemented faultAddToRingBuffer()
- New diagnostics API: faultGetRingBufferDropCount(), faultGetRingBufferEntryCount(), faultGetRingBufferEntry()
- Ring buffer tracks total dropped faults and cycles through entries when full
- Prevents critical faults from being silently discarded under queue pressure
- CLI can now display buffered faults when queue overflowed

INPUT VALIDATION LAYER:
- config_unified.h/cpp: Added configGetIntValidated() and configGetFloatValidated()
- Bounds checking with min/max parameters (negative values skip bounds check)
- Logs warnings when values exceed bounds and clamps to valid range
- Prevents configuration values from causing downstream errors
- Usage: `configGetIntValidated(KEY_SPEED, default, 0, 200)` for 0-200 range

SUMMARY:
- 10+ timeout comparisons now wraparound-safe across millis() rollover
- Fault log can no longer silently lose critical events (ring buffer fallback)
- Configuration values always within expected bounds before use
- System more robust across 49-day uptime threshold

Testing:
- Monitor timeout patterns in motion control (homing, stopping)
- Observe ring buffer in use: `faults show` CLI command will display buffered entries
- Verify config validation: set out-of-bounds values and observe clamping

+++ src/safety.cpp
@@ -40,8 +40,9 @@ void safetyInit() {
 
 void safetyUpdate() {
   uint32_t now = millis();
-  
-  if (now - last_stall_check > SAFETY_STALL_CHECK_INTERVAL_MS) {
+
+  // PHASE 5.1: Wraparound-safe timeout comparison
+  if ((uint32_t)(now - last_stall_check) > SAFETY_STALL_CHECK_INTERVAL_MS) {
     last_stall_check = now;
     
     uint32_t stall_limit_ms = (uint32_t)configGetInt(KEY_STALL_TIMEOUT, SAFETY_MAX_STALL_TIME_MS);
@@ -63,7 +64,8 @@ void safetyUpdate() {
   }
   
   if (alarm_active) {
-    safety_state.fault_duration_ms = millis() - alarm_trigger_time;
+    // PHASE 5.1: Wraparound-safe duration calculation
+    safety_state.fault_duration_ms = (uint32_t)(millis() - alarm_trigger_time);
   }
 }
 
@@ -95,12 +97,13 @@ void safetyTriggerAlarm(const char* reason) {
 
 void safetyResetAlarm() {
   if (!alarm_active) return;
-  
+
   alarm_active = false;
   digitalWrite(SAFETY_ALARM_PIN, LOW);
   safety_state.current_fault = SAFETY_OK;
-  safety_state.fault_duration_ms = millis() - alarm_trigger_time;
-  
+  // PHASE 5.1: Wraparound-safe duration calculation
+  safety_state.fault_duration_ms = (uint32_t)(millis() - alarm_trigger_time);
+
   Serial.printf("[SAFETY] [OK] Alarm reset (Duration: %lu ms)\n", (unsigned long)safety_state.fault_duration_ms);
 }
